Added all functionality
- Yes this commit has too many code changes, this because I programmed beforehand without committing 😅

- Some polishing will definitely follow, like:
- Add feature
- Automagically use the account name, instead of its Account ID
- Tried to implement a tray icon, so a user can switch accounts and see to which account he's logged in, but there's no WinAppSDK native way to do this.
There's a promising-enough library for this: https://github.com/HavenDV/H.NotifyIcon, but it seems to be lacking the ability to add other UI controls than basic Buttons / Commands. So a ListBox is not possible, too bad.
Will yet look into though - feels like a really nice feature to have